Block tests split WIP
